backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
openai>=1.6.0
networkx>=3.2.0
python-dotenv>=1.0.0
sse-starlette>=1.8.0
pydantic>=2.5.0
